[Messaging] Update attributes available for filtering
Related ACR - http://suprem.sec.samsung.net/jira/browse/TWDAPI-205 -
makes it possible to use more attributes to filter
Messages, MessageFolders and MessageConversations.
This commit adds the information about newly supported attributes to
the "Data Filtering and Sorting" guide.
It also adds HTML anchors to the tables with respective message filter
attributes, so that it will be possible to link directly to them.

Signed-off-by: Pawel Wasowski <p.wasowski2@samsung.com>