[NUI] Add guides for Navigation of screens in NUI
NUI provides a system to organize screens and navigate between the
screens by using Page and Navigator.

To organize a formatted full screen page, ContentPage is used.
To show a dialog on the screen, DialogPage is used.

Navigator pushes and pops pages to its LIFO structure.
The peek page, which has been pushed last, is displayed on the screen by
overlapping the previous page.